migrated core llm lib code to sdk
