+++ lib/index.ts
@@ -1,2 +1,3 @@
 export * from './entry'
 export * from './stdin'
+export * from './llm'
